feat: separate `MultiSelect` from `Select` component
